\name{azul-package}
\alias{azul-package}
\alias{azul}
\docType{package}
\title{azul: Thesis-Style Interpretation of Statistical Models and Tables}
\description{
Converts fitted models, hypothesis tests, result tables and raw datasets into
ready-to-paste, manuscript-style interpretation paragraphs following the
Malaysian Journal of Medical Sciences (MJMS) reporting style (Arifin et al.
2016) and SAMPL guidelines. The single entry point is \code{\link{interpret}}.
}
\references{
Arifin WN, et al. Reporting statistical results in medical journals.
Malays J Med Sci. 2016;23(5):1-7.
}
\author{AZul}
